Add tests of offline mode and re-organize tests
