+++ wasm/src/programs/manager/execute.rs
@@ -151,79 +151,3 @@ impl ProgramManager {
         Ok(Transaction::from(transaction))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use wasm_bindgen_test::*;
-    wasm_bindgen_test_configure!(run_in_browser);
-
-    pub const HELLO_PROGRAM: &str = r#"program hello.aleo;
-function main:
-    input r0 as u32.public;
-    input r1 as u32.private;
-    add r0 r1 into r2;
-    output r2 as u32.private;
-"#;
-
-    #[wasm_bindgen_test]
-    async fn test_program_execution_with_cache_and_external_keys() {
-        // Run program locally with cache enabled
-        let mut program_manager = ProgramManager::new();
-        let private_key = PrivateKey::new();
-        let inputs = js_sys::Array::new_with_length(2);
-        inputs.set(0, wasm_bindgen::JsValue::from_str("5u32"));
-        inputs.set(1, wasm_bindgen::JsValue::from_str("5u32"));
-        let result = program_manager
-            .execute_local(private_key.clone(), HELLO_PROGRAM.to_string(), "main".to_string(), inputs, true, None, None)
-            .unwrap();
-        let outputs = result.get_outputs().to_vec();
-        console_log!("outputs: {:?}", outputs);
-        assert_eq!(outputs.len(), 1);
-        assert_eq!(outputs[0], "10u32");
-
-        // Ensure the keys were synthesized in the program run and are in the cache
-        let mut keypair = program_manager.get_cached_keypair("hello.aleo", "main").unwrap();
-
-        // Assert cached keys are used in future program executions
-        let inputs = js_sys::Array::new_with_length(2);
-        inputs.set(0, wasm_bindgen::JsValue::from_str("15u32"));
-        inputs.set(1, wasm_bindgen::JsValue::from_str("5u32"));
-        let result = program_manager
-            .execute_local(private_key.clone(), HELLO_PROGRAM.to_string(), "main".to_string(), inputs, true, None, None)
-            .unwrap();
-
-        // Ensure the output using cached keys is correct
-        let outputs = result.get_outputs().to_vec();
-        console_log!("outputs: {:?}", outputs);
-        assert_eq!(outputs.len(), 1);
-        assert_eq!(outputs[0], "20u32");
-
-        // Assert cached keys are used in future transactions
-        let inputs = js_sys::Array::new_with_length(2);
-        inputs.set(0, wasm_bindgen::JsValue::from_str("15u32"));
-        inputs.set(1, wasm_bindgen::JsValue::from_str("15u32"));
-
-        // Ensure a function execution can be completed via passing external keys to the execute
-        // function
-        let retrieved_proving_key = keypair.proving_key().unwrap();
-        let retrieved_verifying_key = keypair.verifying_key().unwrap();
-        let result = program_manager
-            .execute_local(
-                private_key,
-                HELLO_PROGRAM.to_string(),
-                "main".to_string(),
-                inputs,
-                false,
-                Some(retrieved_proving_key),
-                Some(retrieved_verifying_key),
-            )
-            .unwrap();
-
-        // Ensure the output is correct
-        let outputs = result.get_outputs().to_vec();
-        console_log!("outputs: {:?}", outputs);
-        assert_eq!(outputs.len(), 1);
-        assert_eq!(outputs[0], "30u32");
-    }
-}
